backend/requirements.txt
fastapi
uvicorn
requests
speechrecognition
pywhatkit
numpy
google-generativeai
pyttsx3
python-dotenv